Estruturando projeto e movendo classes para seus arquivos
