+++ backend/lib/challenge.rb
@@ -1,0 +1,9 @@
+require './lib/entities/customer/address'
+require './lib/entities/customer/customer'
+require './lib/entities/membership/membership'
+require './lib/entities/order/order_item'
+require './lib/entities/order/order'
+require './lib/entities/payment/credit_card'
+require './lib/entities/payment/invoice'
+require './lib/entities/payment/payment'
+require './lib/entities/product/product'
