+++ backend/lib/entities/customer/address.rb
@@ -1,0 +1,7 @@
+class Address
+  attr_reader :zipcode
+
+  def initialize(zipcode:)
+    @zipcode = zipcode
+  end
+end
