+++ backend/lib/entities/customer/customer.rb
@@ -1,0 +1,3 @@
+class Customer
+  # you can customize this class by yourself
+end
