+++ backend/lib/entities/membership/membership.rb
@@ -1,0 +1,3 @@
+class Membership
+  # you can customize this class by yourself
+end
