+++ backend/lib/entities/product/product.rb
@@ -1,0 +1,8 @@
+class Product
+  # use type to distinguish each kind of product: physical, book, digital, membership, etc.
+  attr_reader :name, :type
+
+  def initialize(name:, type:)
+    @name, @type = name, type
+  end
+end
